Fixes related to SCN-18-58
The format for the paths and filenames needed to be fixed slightly.
Variations between the two models were tested, and fixed by trial and
error.

In the NOMADS initialization class, the code was over running the
CPLPrintf string buffer, and was causing an issue.  Strings in question
were malloc'ed on the heap.

Fixes #298