gui: add ability to specify output directory
Allow the user to set where the outputs are saved in the GUI.  The
following process is used:

- When an input file is set, the output path is always derived from that
path.
- The user can override the output path after it has been automatically
set
- If a new input file is specified, then the output path is reset

We may want to override the 'reset' path to keep the user defined path,
so it consistent one way or the other.  Currently we aren't sure which
is the more useful method.  We may also want to store the output path in
settings, and load it on startup, but there is no 'correct' way at this
point.

Fixes #369