ninja: fix GDAL 3.0+ calls in GDALGetCenter()
RFC 73[0] introduces a better way to handle axis ordering in
projections, allowing for querying and setting the order in some cases.

This fixes one such case that was found in #383, and applies the same
fix to a few other spots in gdal_util.cpp.  FetchTimeZone() has an
instance that needs to be looked at.  #383 will be re-purposed to track
all of these.

Updates #383