At this point in time, I'm trying to run the WindNinja commands manually from the command line with a simplified case, updating /data/ninjafoam/8/ files as needed to keep up with the changes that were required to get the simplified case to run.
Updated the /data/ninjafoam/8/ files to run all the way up through moveDynamicMesh. Looks like there were very few changes to the moveDynamicMesh stuff, though I had to do a little bit of tweaking to the /constant/fvSolution/ to get it to work, I ended up catching a lot of stuff that I had not updated correctly in /constant/polyMesh/blockMeshDict that I now have correct for OpenFOAM 8. To get the moveDynamicMesh stuff right, I found a useful tutorial in /opt/openfoam8/tutorials/mesh/moveDynamicMesh/ called SnakeRiverCanyon.

I also caught an info line that I forgot to comment out in the OpenFOAM 8 libWindNinja.so applications, the info for those functions were for debugging purposes and were polluting the terminal with unnecessary output.

I also caught a few small things that needed updated in general for the various controlDicts. Because it doesn't have any of the sampling functions in it, controlDict_moveDynamicMesh should be finished, the others still might need a few more tweaks as I get to them in the process.

For Issue #458