Went through the WindNinja code to try to replicate what I've been able to do so far in OpenFOAM 8. With these small changes, and deleting the functions(); lines within /data/ninjafoam/8/system/controlDict_simpleFoam, WindNinja was able to run all the way up till the calls to sample().
Turns out the repeat defining of pszArchive within the loop of NinjaFoam::WriteFoamFiles() was necessary after all, had to put it back in because every few iterations in the loop pszArchive would lose its value. That was what was causing WindNinja to die during calls to NinjaFoam::WriteFoamFiles(). I couldn't track down why pszArchive would lose its value, seems to be related to NinjaFoam::AddBcBlock() as it happens right after writing the epsilon file but not after the T file, but redefining the pszArchive each loop in NinjaFoam::WriteFoamFiles() fixes the problem. With that small change, I was able to get WindNinja to run all the way to simpleFoam.

The change from /system/fvSchemes divSchemes div((nuEff*dev(T(grad(U)))))    Gauss linear; to div((nuEff*dev2(T(grad(U)))))   Gauss linear; has to do with changes in the internal code of the kEpsilon model, a change in the method used to calculate one of the terms. My porous media libWindNinja.so has its own version of the kEpsilon model that went back to the WindNinja OpenFoam 2.2.0 method, as I was trying to replicate it as close as possible back then. But as I didn't see much of a difference, I wanted to try out the current OpenFOAM 8 kEpsilon model methods first.

So this is good, we just need to figure out what the OpenFOAM 8 equivalent to the OpenFOAM 2.2.0 sample function is, updating the /system/controlDict files and calls to sample for the two different versions of OpenFOAM, then we should be good to go.

For Issue #458. [ci skip]