Correcting github actions build dependency cacheing
Looks like I had the paths slightly off, /windninja/scripts/ is the standard current folder location when building on a desktop, but the current folder location for github actions appears to be ${{github.workspace}} which is /home/runner/work/windninja/windninja, so one step above the usual /scripts/ directory. Also, turns out using the hashFiles() wasn't actually necessary for generating the cache key, a simpler non-hashed key should hopefully be good enough.

If this works as intended, it should still cause the commits that use the cache to break, because the other dependencies installed via apt-get are not cached.