Correcting last commit to github actions cacheing
Oops, I forgot to update the PREFIX to the actual values when copying and pasting the ./configure commands for installing poppler, proj, and gdal.

Looks like the ./configure takes a heck of a lot of time as well, not just the regular make commands. So if this works, then the sudo install files made from the cache files should be cached as well as the next step.