Yet another time trying to get the last commit to github actions cacheing to work
Looked like that last time worked till the testing happened, where it looked like somehow the gdal built slightly differently. But I also caught that I didn't return to the starting directory.

Went through and inspected the last full working build commit github actions, to try to see if some intermediate step to download szlib or zlib or libz happened, but I didn't catch anything like that. I did see that trying to track down all the individual files placed all over the place from sudo install for poppler especially, but also for gdal and proj, does not seem easy. Plus it involved editing the PATH variable, not sure that I want to deal with that. So maybe the sudo install step is better to be left outside of the cache after all. I'm also not sure whether library linking is preserved during the cacheing process.

Hopefully this works this time.