Trying to fix github actions
The build is failing during gdal build as it tries to build with sql and odbc, because the github action runner image has changed how it is built with those packages. WindNinja built fine for me on a fresh VMware image of Ubuntu 20.04.

Here is the change in the image that appears to have caused the problem: https://github.com/actions/runner-images/commit/1f4130e15ec1ac0e0861ccbb9beed3d519809dab
That commit/issue refers to this feed here, which appears to be the underlying problem that they were trying to address: https://github.com/microsoft/linux-package-repositories/issues/39

The fix that I am trying here in this commit may not work, sounds like the issue is deeper than just a simple apt install of the missing odbc header files, that it is conflicting files at installation, but I want to try the simple fix before the more complex fix just in case.