Try Again Test github actions cacheing
Oops, forgot to return to base dir for one of the updated folder locations, also forgot that you need a path to the executable, not just a call to the executable to run build_deps.sh.

Looks like it DID start building a new cache though, just from the change in file/folder names within the cache. Not sure if the old files/folders are still saved wherever the old cache was at, but at least this means that updating the cache is as easy as updating folder names of the cache.

Also, I looked back through the past builds and saw that the reason dropping the odbc stuff worked with just the sudo apt update and upgrade, was because odbc is updated with those commands.