one last push to the pointInitialization stationFetch stuff for issue #506, at least for now
updated some of the deprecated gdal functions, but only within pointInitialization::fetchStationData(). The old deprecated gdal functions still seem to be working, and we still need to test this on windows, so I left them in the code for now

that being said, I only updated the deprecated gdal vector reading functions in the pointInitialization::fetchStationData() function, it seemed like there were other places within pointInit still using the deprecated functions, plus who knows where else in the code. In fact, the very getFeature() function that was causing so much trouble for issue #506 is still being used in some of the other parts of the pointInitialization code, going to need to watch out to see whether they cause problems in the future