updated some of the same deprecated gdal ogr functions found in pointInit fetchMetaData() as I updated in pointInit fetchStationData(), for issue #506
note that this is STILL not an update to all deprecated gdal or gdal ogr functions within pointInit or within windninja, turns out that there's enough going on with various deprecated since gdal 2.0 gdal ogr functions within windninja that it needs to be its own separate issue and separate work. In addition, see the weird stuff mixing and maybe even doing gdal ogr stuff twice within pointInit fetchWxStationID(), there's a mixing of C and C++ style gdal functions there and a decision which to pick and which to drop needs to be made, so yes any more updating of deprecated gdal stuff probably needs to be done in a separate issue

still, I wanted to at least do the small update to the deprecated functions where I was working when dealing with this issue, to make sure it wasn't something else going on. Turns out that the only problem causing the original issue was the supposedly NON-deprecated OGR_L_GetFeature() function that we replaced with OGR_L_GetNextFeature(), which wasn't occuring outside fetchMetaData() and fetchStationData() within pointInit. But might as well keep the small deprecated function update