revisit of issue #458, foam 8 vs 2.2.0 stuff
small cleanup fix to openfoam 2.2.0 case directory nut file, makes paraview complain a ton less, also makes it more compatible going back and forth between versions of OpenFOAM with the same case generated by both versions of OpenFOAM. Also small cleanup to the openfoam 2.2.0 inlet BC applications to actually print out the uDirection when running anything foam as that also helps with compatibility going back between versions of OpenFOAM, also why print it within WindNinja if the OpenFOAM BC is gonna drop it whenever it is called, made sense to make it consistent.

was going to do an additional commit to maybe cleanup the BC function names, to also make the cases more compatible back and forth, but I was pleased to find that the current methodology already has the compatible and same BC function names that I was thinking of.

with these small changes, I was able to run both versions of OpenFOAM on the same case, and open both cases with both versions of OpenFOAM/paraview. Well had to drop the functions part of the controlDict for the foam 8 case, but that was just commenting stuff out, not adding stuff back in.

for issue #458