step back a bit last fix to pointInput gui, for the pointInput gui timezone fix
looks like even with this rollback, the general behavior of the pointInput gui station time selection is on a per file basis, rather than an across all selected files basis. In particular, it seems to choose the times and timestep based off of the last file selected rather than across all the files, and when deselecting a file it chooses the times and timestep based off of the last file or second to last file in the queue rather than a true across all files basis. This deselecting behavior makes it imitate working on an across all selected files basis, but it is not quite right. Whatever the case, the current behavior isn't picking the time range using all the selected files at once like is done in wxModel

the differing file types check IS working on an across all selected files basis though, the red x disallowing WindNinja runs till only files of the same type are selected is correct. Implies overall  proper behavior with directStationTraffic(), just some parts of directStationTraffic() probably need migrated to the function calling, so some reorganization of the call order

I did notice a few other strange things, this case seems to be a single timestep, that is NOT a latest time simulation, and the fix that I've applied so far is essentially treating that type of time as a multiple time step. Should an additional type of time for this case be created instead, or should the current fix be continued, and along each step of the code make sure it correctly handles if it is a multi-timestep time but as a single time? For example, WindNinja (not the gui), let me run from the gui the test case files of 3 stations of a single time, with 2 timesteps. It ran the first timestep, as it did when specifying the timestep to be 1, then it choked in an infinite loop trying to do a second timestep. This seems like it should be disallowed or checked for somewhere in the code, at the solver level not just the gui

in addition, I did my test case by copying and pasting the test files, same names new appendix same stations, and the gui let me select each of them together even with the same. Then the solver failed to run with a warning that separate station files had the same station name

I guess I'm wondering if at some time, maybe some refactoring might be done to try to make the gui use the same error checks as the solver somehow, make the gui run more solver code instead of doing its own checks for the pointInitialization stuff. Does seem like a bunch of extra work though, just a thought when I saw this

For issue #516