pointInput gui timezone fix
added check to better handle if single time is run as a multi time, but with more than one timestep, WindNinja needs to run it as a single time

fix has some problems. At first, just tried an error to kick it out if the condition existed, but as this pointInitialization::getTimeList() function is NOT called within a normal try/catch like the actual solve() stuff, it resulted in WindNinja just closing with the error rather than just stopping and allowing the restart of the simulation with the message. Hence I commented that out, might still want it but not sure yet. So instead I manually just define the created timeList to just be a single time, with warning letting the user know it is happening

so not ideal, nTimeSteps isn't technically updated elsewhere, just ignored in its last use. So need to find another fix that is better

for issue #516