pointInput gui timezone fix
stepped back the last fix, mostly just wanted it in the record in case need to come back to it. Put the fix into a better spot, one that is called by both gui and cli, and that let's the solver stop and let the user start a new simulation updating the variables, with a good error message of what needs fixed

noticed that this pointInitialization::validateTimeData() function the fix goes into has some issues with the method used to setup and pass around the messages, and did the fix the proper way. Probably need to clean up this function for the other messages while I'm at it

for issue #516