cleaned up the pointInit cli config options stuff a bit, to make it less confusing when missing an option
walked back the option dependency for wx_station_filename that was recently added that was confusing. New verify_option_set(vm, "wx_station_filename"); line SHOULD fix this, but is not the most ideal message? But there wasn't really quite anything good to tie to as an option dependency? "initialization_method" doesn't quite work because want to specifically give a message that it is a "pointInitialization" dependency. Anyhow, this verify_option_set(vm, "wx_station_filename"); fix should work for now.