found and fixed some more stuff in ascii2png, specifically when running turbulence output for another case, missoula_valley, the band of noData vals from converting coordinate systems showed up as the min color instead of as transparent
turned out to be caused by using 0 to 255 for the range of colors, leaving no color available for noData vals, when gdal warp was called it set the noData vals to the 0 color. Setting the histogram binned values of colors from 1 to 255, and manually setting noData vals to 0 fixed the issue. This was also tested to work for incoming datasets already filled with patches of noData values, not just for noData values introduced from gdal warp.

found a new quirky issue where the range of colors and values for the colors differs somehow after gdal warp, leading to small patches of empty plot values for the max values of the colorbar range, still happens with this fix, may have to generate the color tables after instead of before the gdal warp to fix it