adjusted ascii2png to no longer need intermediate ascii grid output, instead gdalwarp is done directly on the already required intermediate tiff file. Also made it possible to keep the intermediate tiff file, though only the new gdalwarped form. Also added an option to the cli for keeping the intermediate tiff file/files when writing turbulence kml/kmz outputs. Also a small fix to cli options to make the "turbulence_output_flag" require google output (kml/kmz output).
originally my goal was to set ascii2png up to pick and choose to do only one png or tiff output, no intermediate datasets, just do a single gdal driver with an if statement to swap back and forth between drivers. But it turns out that the reason ascii2png was always doing png output with both png and tiff gdal drivers and intermediate files was because the gdal png driver does not support the gdal create() function, but only the gdal createCopy() function. So the standard way to do a png output in gdal is to create a tiff file then make a copy of the tiff file as a png file using createCopy(). Silly, but apparently that's been the standard way in the past, and still seems to be the standard way with our current version of gdal. See https://gis.stackexchange.com/questions/132298/gdal-c-api-how-to-create-png-or-jpeg-from-scratch.
As the tiff file was already being created alongside the png to do png output (kml/kmz output), but the tiff file is now required in addition to kml/kmz output, seemed best to just provide the option to keep the intermediate tiff file and improve its creation methods rather than making an additional ascii2tiff function.

Important side note, turns out the remaining small empty patchy spots in the ascii2png output from the past fix to the ascii2png colormapping was NOT related to gdalwarping the image, turns out that the png and intermediate tiff files before and after gdalwarp actually had no holes in them. This can be seen when viewing them as intermediate outputs before they are stuffed into the kmz file. The holes turned out to only be something that happens when viewing the end result kmz files in google earth. The holes go away in google earth when deselecting the google earth "3D Buildings" layer from the view. In other words, turns out that last fix to ascii2png colormap was fully correct and not still missing something.