preparing ascii2png() to work with different numbers of colors and coloring style. changed the code method for the png legend
The old method for the png legend involved drawing 20 color lines per cbar color box at pixel y positions estimated from percent image height values. Had a lot of hard to track adjustment methods as well. In addition, the conversion from percent image height values to pixel y positions, int on a double, caused some lines to be repeat lines, redrawing the same line in the same spot.

The new method involves using the textHeight and the estimated number of padded pixels when drawing the text of numbers, to calculate the number of pixel y positions to draw for each color box. Also some reworking of a few of the positioning methods to drop out some of the quirky adjustments.

The new method is not perfect in that DrawString() calls DrawLine() outputting the exact number of pixels as is specified by textHeight (well, bigger textHeight could add a pixel still in the precall to DrawLine() and other Draw() methods), but calls to DrawArc() uses round in such a way as to randomly pad up to 1 pixel above and below, in addition to always padding 1 pixel above and below to the value. The methods in EasyBMP have just enough quirkiness to them that it isn't easy to calculate the effective textHeight output from calls to DrawString() from the input textHeight, but the estimates that I used seem to be close enough