made new ascii2png color ramp methods available by the command line for the turbulence kml/kmz outputs
the way it was done is still a bit quirky, in that it expects a default use of the color ramp going from min to max with 4 color breaks as we usually do, and you override the default use with 3 or 4 specific color break values to use. To make such stuff as inputs also leads to quirkiness in different variable names and copies of variables at various levels, the messages were kept at the ascii2png level, but it works great for what was intended.

Still feels way too rigid to me, all the extra arguments tracking everywhere in ascii.cpp also made me think yet again that this stuff should be in its own class. Didn't realize till afterwards that my debugging functionality of going from min to max with 4 colorbreaks in different spots than the default min to max use, or 3 colorbreaks in the same way, is not yet available in the command line, though it can still be done manually through hard coded values.

improved some of the checks a bit as well. Did a LOT of testing on the cli inputs, seems like they should work rigorously as intended. But again, still kind of rigid in how it all came together, might not be fun to edit for future use.

note that while the color legend definitely lets the inputs have negative break values, and plot looks fine, still might be a chance that the coloring gets off for such an edge case. Might be good to do some additional checking of the methods if ever running into use where the plotted values can go negative.